pyproject.toml
[tool.poetry]
name = "ffmpeg-liteweight-video-compressor"
version = "0.1.0"
description = "UI for target folder and start convert video file to hevc/h265 codec"
authors = ["Cheong Gwen Yit"]

[tool.poetry.dependencies]
python = "^3.10"
ffmpeg-python = "^0.2.0"
get-video-properties = "^0.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
